% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame  A data.frame that contains the weight and value of each
element.}

\item{W}{integer  An integer that shows the limit for the weight that can be
carried by the knapsack.}
}
\description{
This function like two other functions
  \code{\link{brute_force_knapsack}} and \code{\link{greedy_knapsack}} is one
  of the aproaches for solving knapsack problem
  (\url{https://en.wikipedia.org/wiki/Knapsack_problem}). The pseudocode for
  this algorithm can be found
  here:\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
\examples{
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)
# $value
# 16770
# $elements
# 5 8

}
