% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{knapsack_objects}
\format{An object of class \code{data.frame} with 2000 rows and 2 columns.}
\usage{
knapsack_objects
}
\description{
knapsack_objects
}
\keyword{datasets}
