% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlab06.R
\docType{package}
\name{Rlab06}
\alias{Rlab06}
\alias{Rlab06-package}
\title{Knapsack Package}
\description{
The package will contain three different functions for solving the knapsack problem.
}
\author{
Group4 \email{yusur.mutair@gamil.com}
}
